** hetero -> default with uniform weights if not specified
